Changed reducer codebase to suite RTK
