+++ 18-Redux-intro/src/features/accounts/accountSlice.js
@@ -1,75 +1,49 @@
-const initialStateAccount = {
+import { createSlice } from '@reduxjs/toolkit'
+
+const initialState = {
   balance: 0,
   loan: 0,
   loanPurpose: '',
   isLoading: false,
 }
 
-export default function accountReducer(state = initialStateAccount, action) {
-  switch (action.type) {
-    case 'account/deposit':
-      return {
-        ...state,
-        balance: state.balance + action.payload,
-        isLoading: false,
-      }
+const accountSlice = createSlice({
+  name: 'account',
+  initialState,
+  reducers: {
+    deposit(state, action) {
+      state.balance += action.payload
+    },
 
-    case 'account/withdraw':
-      return { ...state, balance: state.balance - action.payload }
+    withdraw(state, action) {
+      state.balance -= action.payload
+    },
 
-    case 'account/requestLoan':
-      if (state.loan > 0) return state
-      return {
-        ...state,
-        loan: action.payload.amount,
-        balance: state.balance + action.payload.amount,
-        loanPurpose: action.payload.purpose,
-      }
+    requestLoan: {
+      // method to accept multiple arguments from dispatched actions
+      prepare(amount, purpose) {
+        return {
+          payload: { amount, purpose },
+        }
+      },
 
-    case 'account/payLoan':
-      return {
-        ...state,
-        loan: 0,
-        loanPurpose: '',
-        balance: state.balance - state.loan,
-      }
+      reducer(state, action) {
+        if (state.loan > 0) return
 
-    case 'account/convertingCurrency':
-      return { ...state, isLoading: true }
+        state.balance += action.payload.amount
+        state.loan = action.payload.amount
+        state.loanPurpose = action.payload.purpose
+      },
+    },
 
-    default:
-      return state
-  }
-}
+    payLoan(state, action) {
+      state.balance -= state.loan
+      state.loanPurpose = ''
+      state.loan = 0
+    },
+  },
+})
 
-export function deposit(amount, currency) {
-  if (currency === 'USD') return { type: 'account/deposit', payload: amount }
+export const { deposit, withdraw, requestLoan, payLoan } = accountSlice.actions
 
-  // middleware function
-  return async function (dispatch, getState) {
-    dispatch({ type: 'account/convertingCurrency' })
-
-    const res = await fetch(
-      `https://api.frankfurter.app/latest?amount=${amount}&from=${currency}&to=USD`
-    )
-    const data = await res.json()
-    const convertedAmt = data.rates.USD
-
-    dispatch({ type: 'account/deposit', payload: convertedAmt })
-  }
-}
-
-export function withdraw(amount) {
-  return { type: 'account/withdraw', payload: amount }
-}
-
-export function requestLoan(amount, purpose) {
-  return {
-    type: 'account/requestLoan',
-    payload: { amount, purpose },
-  }
-}
-
-export function payLoan() {
-  return { type: 'account/payLoan' }
-}
+export default accountSlice.reducer
